Start using NSInteger and NSUInteger
These have been introduced in Leopard and default to int and unsigned int.
In 64-bit Snow Leopard they are long and unsigned long.  This caused issues
with the getRectsBeingDrawn message which needs a pointer to a NSInteger
(long on 64-bit!) but we passed in an integer.  Surprisingly this problem
was visible when compiling with -O0 (segfault), but *not* when compiling
with -O1.  Other messages were NSInteger is now needed have also been
adapted.

Since NSInteger and NSUInteger are not available on Tiger, a define
has been added to add typedefs for these when they have not been defined
by the system headers.

+++ gdk/quartz/gdkcursor-quartz.c
@@ -84,7 +84,7 @@ create_builtin_cursor (GdkCursorType cursor_type)
 {
   GdkCursor *cursor;
   NSBitmapImageRep *bitmap_rep;
-  gint mask_width, mask_height;
+  NSInteger mask_width, mask_height;
   gint src_width, src_height;
   gint dst_stride;
   const guchar *mask_start, *src_start;
@@ -249,7 +249,7 @@ gdk_cursor_new_from_pixmap (GdkPixmap      *source,
   NSImage *image;
   NSCursor *nscursor;
   GdkCursor *cursor;
-  gint width, height;
+  NSInteger width, height;
   gint tmp_x, tmp_y;
   guchar *dst_data, *mask_data, *src_data;
   guchar *mask_start, *src_start;
